Support webpack 4
